Spilt the files according to their own roles
- ViewAccessibilityV2.cs / ViewAccessibilityRoleV2.cs / ViewAccessibilityStatesV2.cs
