+++ src/Tizen.NUI/src/public/BaseComponents/ViewAccessibilityRoleV2.cs
@@ -1,0 +1,171 @@
+/*
+ * Copyright(c) 2025 Samsung Electronics Co., Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+using System;
+using System.ComponentModel;
+using Tizen.NUI;
+
+namespace Tizen.NUI.BaseComponents
+{
+    /// <summary>
+    /// Specifies the role of an accessible object.
+    /// </summary>
+    [EditorBrowsable(EditorBrowsableState.Never)]
+    public enum AccessibilityRoleV2
+    {
+        /// <summary>
+        /// A slider
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Adjustable = 200,
+        /// <summary>
+        /// An alert
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Alert,
+        /// <summary>
+        /// A button
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Button,
+        /// <summary>
+        /// A check box
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        CheckBox,
+        /// <summary>
+        /// A combo box
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        ComboBox,
+        /// <summary>
+        /// A container
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Container,
+        /// <summary>
+        /// A dialog
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Dialog,
+        /// <summary>
+        /// An entry
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Entry,
+        /// <summary>
+        /// A header
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Header,
+        /// <summary>
+        /// An Image
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Image,
+        /// <summary>
+        /// A link
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Link,
+        /// <summary>
+        /// A list
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        List,
+        /// <summary>
+        /// An item of the list
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        ListItem,
+        /// <summary>
+        /// A menu
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Menu,
+        /// <summary>
+        /// A menu bar
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        MenuBar,
+        /// <summary>
+        /// An item of the menu
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        MenuItem,
+        /// <summary>
+        /// None
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        None,
+        /// <summary>
+        /// A password text
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        PasswordText,
+        /// <summary>
+        /// A popup menu
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        PopupMenu,
+        /// <summary>
+        /// A progress bar
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        ProgressBar,
+        /// <summary>
+        /// A radio button
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        RadioButton,
+        /// <summary>
+        /// A scroll bar
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        ScrollBar,
+        /// <summary>
+        /// A spin button
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        SpinButton,
+        /// <summary>
+        /// A tab
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Tab,
+        /// <summary>
+        /// A tab list
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        TabList,
+        /// <summary>
+        /// A text
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Text,
+        /// <summary>
+        /// A toggle button
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        ToggleButton,
+        /// <summary>
+        /// a tool bar
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        Toolbar,
+    }
+}
