+++ src/Tizen.NUI/src/public/BaseComponents/ViewAccessibilityV2.cs
@@ -17,251 +17,10 @@
 
 using System;
 using System.ComponentModel;
-using System.Runtime.InteropServices;
 using Tizen.NUI;
 
 namespace Tizen.NUI.BaseComponents
 {
-    /// <summary>
-    /// AccessibilityStateV2 is an enumeration that represents the states of the view to send to accessibility service.
-    /// </summary>
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public enum AccessibilityStateV2
-    {
-        /// <summary>
-        /// Indicates whether the view is enabled or not.
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Enabled = 0,
-        /// <summary>
-        /// Indicates whether a selectable element is currently selected or not.
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Selected,
-        /// <summary>
-        /// Indicates the state of a checkable element.
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Checked,
-        /// <summary>
-        /// Indicates whether an element is currently busy or not.
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Busy,
-        /// <summary>
-        /// Indicates whether an expandable element is currently expanded or collapsed.
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Expanded
-    }
-
-    /// <summary>
-    /// The AccessibilityStatesV2 structure represents a set of states of the view to communicate to accessibility service.
-    /// </summary>
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public struct AccessibilityStatesV2
-    {
-        const uint c_on = 1;
-        private uint _bitMask;
-
-        /// <summary>
-        /// Gets or sets the value of the specified accessibility state flag using bitmask operations.
-        /// </summary>
-        /// <param name="state">The <see cref="AccessibilityStateV2"/> enum value representing the accessibility state to check or modify.</param>
-        /// <returns>
-        /// <c>true</c> if the specified state flag is set in the bitmask; otherwise, <c>false</c>.
-        /// </returns>
-        /// <remarks>
-        /// This indexer uses bitwise operations to efficiently store and retrieve multiple boolean state flags
-        /// in a single integer field (_bitMask). Each state corresponds to a specific bit position determined
-        /// by its enum value.
-        /// </remarks>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public bool this[AccessibilityStateV2 state]
-        {
-            get
-            {
-                return Convert.ToBoolean(_bitMask & (c_on << (int)state));
-            }
-            set
-            {
-                if (value)
-                {
-                    _bitMask |= (c_on << (int)state);
-                }
-                else
-                {
-                    _bitMask &= ~(c_on << (int)state);
-                }
-            }
-        }
-
-        AccessibilityStatesV2(int states)
-        {
-            _bitMask = (uint)states;
-        }
-
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static explicit operator int(AccessibilityStatesV2 state)
-        {
-            return (int)state._bitMask;
-        }
-
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static explicit operator AccessibilityStatesV2(int states)
-        {
-            return new AccessibilityStatesV2(states);
-        }
-    }
-
-    /// <summary>
-    /// Specifies the role of an accessible object.
-    /// </summary>
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public enum AccessibilityRoleV2
-    {
-        /// <summary>
-        /// A slider
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Adjustable = 200,
-        /// <summary>
-        /// An alert
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Alert,
-        /// <summary>
-        /// A button
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Button,
-        /// <summary>
-        /// A check box
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        CheckBox,
-        /// <summary>
-        /// A combo box
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        ComboBox,
-        /// <summary>
-        /// A container
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Container,
-        /// <summary>
-        /// A dialog
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Dialog,
-        /// <summary>
-        /// An entry
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Entry,
-        /// <summary>
-        /// A header
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Header,
-        /// <summary>
-        /// An Image
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Image,
-        /// <summary>
-        /// A link
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Link,
-        /// <summary>
-        /// A list
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        List,
-        /// <summary>
-        /// An item of the list
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        ListItem,
-        /// <summary>
-        /// A menu
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Menu,
-        /// <summary>
-        /// A menu bar
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        MenuBar,
-        /// <summary>
-        /// An item of the menu
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        MenuItem,
-        /// <summary>
-        /// None
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        None,
-        /// <summary>
-        /// A password text
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        PasswordText,
-        /// <summary>
-        /// A popup menu
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        PopupMenu,
-        /// <summary>
-        /// A progress bar
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        ProgressBar,
-        /// <summary>
-        /// A radio button
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        RadioButton,
-        /// <summary>
-        /// A scroll bar
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        ScrollBar,
-        /// <summary>
-        /// A spin button
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        SpinButton,
-        /// <summary>
-        /// A tab
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Tab,
-        /// <summary>
-        /// A tab list
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        TabList,
-        /// <summary>
-        /// A text
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Text,
-        /// <summary>
-        /// A toggle button
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        ToggleButton,
-        /// <summary>
-        /// a tool bar
-        /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        Toolbar,
-    }
-
     /// <summary>
     /// View is the base class for all views.
     /// </summary>
